+++ jsx-dev-runtime.js
@@ -2,8 +2,8 @@
 
 const { Fragment, createElement } = require('./src/htmx/Jsx');
 
-function jsxDEV(type, properties) {
-    return createElement(type, properties);
+function jsxDEV(type, properties, key) {
+    return createElement(type, properties, key);
 }
 
 module.exports = { Fragment, jsxDEV };